Exercise 4.21

+++ 4cd/controllers/blogs.js
@@ -11,7 +11,7 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
+        return response.status(401).json({error: 'token invalid'})
     }
     const user = await User.findById(decodedToken.id)
 
@@ -40,6 +40,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({error: 'token invalid'})
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if(!blog) {
+        return response.status(404).json({error: 'blog not found'})
+    }
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+        return response.status(401).json({error: 'cannot delete, you are not an author'})
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
